backend/requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
sqlalchemy
alembic
python-jose
passlib
aiosmtplib
pydantic
pytest
httpx
openai
python-dotenv
pydantic_settings
python-multipart
anthropic
aiofiles
supabase
websockets
pillow
bcrypt
cryptography